enable versbose logging in JsDebugger

+++ src/src/com/tns/JsDebugger.java
@@ -48,6 +48,8 @@ public class JsDebugger
 
 	private String baseDebugFilesDir;
 	
+	private static final boolean LOGGING_ENABLED = true;
+	
 	private static final String portEnvInputFile = "envDebug.in";
 	
 	private static final String portEnvOutputFile = "envDebug.out";
@@ -316,7 +318,7 @@ public class JsDebugger
 
 	int getDebuggerPortFromEnvironment()
 	{
-		if (Platform.IsLogEnabled)
+		if (Platform.IsLogEnabled || LOGGING_ENABLED)
 		{
 			Log.d(Platform.DEFAULT_LOG_TAG, "getDebuggerPortFromEnvironment");
 		}
@@ -325,9 +327,9 @@ public class JsDebugger
 		
 		boolean shouldEnableDebuggingFlag = shouldEnableDebugging(context);
 		
-		if (Platform.IsLogEnabled)
+		if (Platform.IsLogEnabled || LOGGING_ENABLED)
 		{
-			Log.d(Platform.DEFAULT_LOG_TAG, "getDebuggerPortFromEnvironment:: shouldEnableDebuggingFlag=" + shouldEnableDebuggingFlag);		
+			Log.d(Platform.DEFAULT_LOG_TAG, "getDebuggerPortFromEnvironment: shouldEnableDebuggingFlag: " + shouldEnableDebuggingFlag);		
 		}
 
 		if (shouldEnableDebuggingFlag)
@@ -339,6 +341,11 @@ public class JsDebugger
 				w = new OutputStreamWriter(new FileOutputStream(envOutFile, false));
 				String currentPID = "PID=" + android.os.Process.myPid() + "\n";
 				w.write(currentPID);
+				
+				if (Platform.IsLogEnabled || LOGGING_ENABLED)
+				{
+					Log.d(Platform.DEFAULT_LOG_TAG, "getDebuggerPortFromEnvironment: writting PID: " + currentPID + " to envOutFile: " + envOutFile.getAbsolutePath());		
+				}
 			}
 			catch (IOException e1)
 			{
@@ -362,14 +369,13 @@ public class JsDebugger
 			
 			boolean shouldDebugBreakFlag = shouldDebugBreak(context); 
 			
-			if (Platform.IsLogEnabled)
+			if (Platform.IsLogEnabled || LOGGING_ENABLED)
 			{
 				Log.d(Platform.DEFAULT_LOG_TAG, "shouldDebugBreakFlag=" + shouldDebugBreakFlag);
 			}
 			
 			if (shouldDebugBreakFlag)
 			{
-				
 				try
 				{
 					Thread.sleep(3 * 1000);
@@ -384,9 +390,9 @@ public class JsDebugger
 			
 			boolean envInFileFlag = envInFile.exists();
 			
-			if (Platform.IsLogEnabled)
+			if (Platform.IsLogEnabled || LOGGING_ENABLED)
 			{
-				Log.d(Platform.DEFAULT_LOG_TAG, "envInFileFlag=" + envInFileFlag);
+				Log.d(Platform.DEFAULT_LOG_TAG, "envInFile exists: " + envInFileFlag);
 			}
 			
 			if (envInFileFlag)
@@ -411,6 +417,12 @@ public class JsDebugger
 					String strLocalPort = "PORT=" + localPort + "\n";
 					w.write(strLocalPort);
 					port = currentPort = localPort;
+					
+					if (Platform.IsLogEnabled || LOGGING_ENABLED)
+					{
+						Log.d(Platform.DEFAULT_LOG_TAG, "Debugger enabled on port: " + port);
+					}
+					
 					//
 					enable();
 					debugBreak();
@@ -451,7 +463,7 @@ public class JsDebugger
 			}
 		}
 		
-		if (Platform.IsLogEnabled)
+		if (Platform.IsLogEnabled || LOGGING_ENABLED)
 		{
 			Log.d(Platform.DEFAULT_LOG_TAG, "Debug port=" + port);
 		}
@@ -570,6 +582,11 @@ public class JsDebugger
 		}
 
 		boolean shouldEnableDebugging = ((flags & ApplicationInfo.FLAG_DEBUGGABLE) != 0);
+		
+		if (Platform.IsLogEnabled || LOGGING_ENABLED)
+		{
+			Log.d(Platform.DEFAULT_LOG_TAG, "shouldEnableDebugging: " + shouldEnableDebugging);
+		}
 		return shouldEnableDebugging;
 	}
 	
@@ -585,6 +602,10 @@ public class JsDebugger
 		
 		if (!shouldEnableDebugging(context))
 		{
+			if (Platform.IsLogEnabled || LOGGING_ENABLED)
+			{
+				Log.d(Platform.DEFAULT_LOG_TAG, "shouldDebugBreak false. Debugging not enabled");
+			}
 			shouldDebugBreakFlag = false;
 			return false;
 		}
@@ -592,11 +613,19 @@ public class JsDebugger
 		File debugBreakFile = new File(baseDebugFilesDir, DEBUG_BREAK_FILENAME);
 		if (debugBreakFile.exists())
 		{
+			if (Platform.IsLogEnabled || LOGGING_ENABLED)
+			{
+				Log.d(Platform.DEFAULT_LOG_TAG, "shouldDebugBreak: Debug-brk file found. Debugger should break");
+			}
 			debugBreakFile.delete();
 			shouldDebugBreakFlag = true;
 			return true;
 		}
 		
+		if (Platform.IsLogEnabled || LOGGING_ENABLED)
+		{
+			Log.d(Platform.DEFAULT_LOG_TAG, "shouldDebugBreak: Debugger should not break");
+		}
 		shouldDebugBreakFlag = false;
 		return false;
 	}
